requirements.txt
fastapi
httpx
beautifulsoup4
sqlalchemy
pydantic
redis
aiofiles
python-dotenv
